requirements.txt
aiogram
colorama
instaloader
mutagen
pillow
python-dotenv
yt-dlp